// src/Storefront.WebApp/app/src/workplace/enums/transaction/txnStatus.enum.ts
export enum TxnStatusEnum {
  New = 'New',
  Loaded = 'Loaded',
  LoadError = 'LoadError',
  Prepared = 'Prepared',
  Finished = 'Finished',
  Rejected = 'Rejected',
  Closed = 'Closed',
  PreProcessing = 'PreProcessing',
  Shaded = 'Shaded',
  PreAppliedAsFin = 'PreAppliedAsFin',
  Undefined = 'Undefined'
}